requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.3.0
python-docx==1.1.0
markdown==3.5.1
camelot-py[cv]==1.0.0
opencv-python-headless==4.11.0.86
numpy==1.26.4
Pillow==10.4.0
tabula-py==2.10.0
jpype1==1.5.2 